+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+4.1.0
+=====
+
+*   Removed automatically fetching attributes from the request (less magic is better).
+*   Add option to set tree parameters when rendering a tree.
+
+
 4.0.1
 =====